requirements.txt
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
groq==0.37.1
schedule==1.2.0
PyQt5==5.15.9
PyQt5-sip==12.13.0
